refactor: encapsulated tax overview into its own container, similar to TayPayments and TaxDetails
